PTSD-764. Allow for frame indices to be signed.

+++ src/basecaller/basecaller/analyzer/BatchAnalyzer.cpp
@@ -173,8 +173,8 @@ BatchAnalyzer::OutputType BatchAnalyzer::operator()(const TraceBatch<int16_t>& t
 BatchAnalyzer::OutputType FixedModelBatchAnalyzer::AnalyzeImpl(const TraceBatch<int16_t>& tbatch)
 {
     auto mode = AnalysisProfiler::Mode::REPORT;
-    if (tbatch.Metadata().FirstFrame() < tbatch.NumFrames()*10+1) mode = AnalysisProfiler::Mode::OBSERVE;
-    if (tbatch.Metadata().FirstFrame() < tbatch.NumFrames()*2+1)  mode = AnalysisProfiler::Mode::IGNORE;
+    if (tbatch.Metadata().FirstFrame() < static_cast<int32_t>(tbatch.NumFrames())*10+1) mode = AnalysisProfiler::Mode::OBSERVE;
+    if (tbatch.Metadata().FirstFrame() < static_cast<int32_t>(tbatch.NumFrames())*2+1)  mode = AnalysisProfiler::Mode::IGNORE;
     AnalysisProfiler profiler(mode, 3.0, 100.0);
 
     auto baselineProfile = profiler.CreateScopedProfiler(AnalysisStages::Baseline);
@@ -224,7 +224,7 @@ BatchAnalyzer::OutputType SingleEstimateBatchAnalyzer::AnalyzeImpl(const TraceBa
     auto baselinedTraces = std::move(baselinedTracesAndMetrics.first);
     auto baselinerMetrics = std::move(baselinedTracesAndMetrics.second);
 
-    if (!isModelInitialized_ && tbatch.GetMeta().FirstFrame() > baseliner_->StartupLatency())
+    if (!isModelInitialized_ && tbatch.GetMeta().FirstFrame() > static_cast<int32_t>(baseliner_->StartupLatency()))
     {
         // Run data through the DME until we get our first real estimate, at which point we
         // stop using the DME and just keep that model forever.
@@ -309,7 +309,7 @@ DynamicEstimateBatchAnalyzer::AnalyzeImpl(const Data::TraceBatch<int16_t>& tbatc
           roundToChunkMultiple(nFramesBaselinerStartUp)
         + roundToChunkMultiple(nFramesDmeStartUp);
     auto mode = AnalysisProfiler::Mode::IGNORE;
-    if (tbatch.Metadata().FirstFrame() > startupLatency)
+    if (tbatch.Metadata().FirstFrame() > static_cast<int32_t>(startupLatency))
     {
         auto framesSince = tbatch.Metadata().FirstFrame() - startupLatency;
         if (framesSince > 2*minFramesForDme)  mode = AnalysisProfiler::Mode::OBSERVE;
@@ -330,7 +330,7 @@ DynamicEstimateBatchAnalyzer::AnalyzeImpl(const Data::TraceBatch<int16_t>& tbatc
     // innacuracies, and we shouldn't run the downstream filters either because
     // the models aren't even sensibly initialized until the first time we hand
     // data to the DME
-    if (baselinedTraces.GetMeta().FirstFrame() < nFramesBaselinerStartUp + poolDmeDelayFrames_)
+    if (baselinedTraces.GetMeta().FirstFrame() < static_cast<int32_t>(nFramesBaselinerStartUp + poolDmeDelayFrames_))
     {
         auto emptyLabels = frameLabeler_->EmptyLabelsBatch(std::move(baselinedTraces));
         auto emptyPulsesAndMetrics = pulseAccumulator_->EmptyPulseBatch(emptyLabels.Metadata(),
